fix: 修复小程序模式合盘选择不显示的问题
